refactor(web/engine): embedded adjustments, defaultOutput split
