+++ web/source/kmwutils.ts
@@ -4,6 +4,8 @@
 /// <reference path="kmwdevice.ts" />
 // Includes the DOM utils, since our UI modules need access to certain methods here.
 /// <reference path="dom/utils.ts" />
+// Include other important utils
+/// <reference path="utils/deepCopy.ts" />
 
 namespace com.keyman {
   class DOMEventTracking {
@@ -909,30 +911,6 @@ namespace com.keyman {
       }
 
       return ''+item;
-    }
-
-    /**
-     * Function     deepCopy
-     * Scope        Private
-     * @param       {Object}      p           object to copy
-     * @param       {Array=}      c0          array member being copied
-     * @return      {Object}                  clone ('deep copy') of object
-     * Description  Makes an actual copy (not a reference) of an object, copying simple members,
-     *              arrays and member objects but not functions, so use with care!
-     */
-    deepCopy<T>(p:T, c0?): T {
-      var c = c0 || {};
-      for (var i in p) {
-        if(typeof p[i] === 'object') {
-          c[i] = (p[i].constructor === Array ) ? [] : {};
-          this.deepCopy(p[i],c[i]);
-        }
-        else {
-          c[i] = p[i];
-        }
-      }
-
-      return c;
     }
 
     /**
